41: replace the big `tests/tests.py` module with several newly-created modules within the `tests/api/` folder
